refactor: 소속 패키지 변경, 어노테이션 수정
