+++ src/main/java/beachcombing/backend/domain/store/Store.java
@@ -1,16 +1,11 @@
-package beachcombing.backend.domain.giftcard.domain;
+package beachcombing.backend.domain.store;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 @Entity
 @Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Store {
 
     @Id
